requirements.txt
requests==2.32.3
jsonschema==4.23.0
pywinauto==0.6.9
pyautogui==0.9.54
pillow==10.4.0
keyboard==0.13.5
pytest==8.3.2
